Remove local_rebuild from Build
